# backend/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: StudyApp API
  version: "1.0.0"
  description: Scan-to-Quiz educational platform API
servers:
  - url: http://localhost:8080
    description: Development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - VALIDATION_ERROR
            - CONFLICT
            - RATE_LIMIT_EXCEEDED
            - FEATURE_NOT_AVAILABLE
            - SCAN_LIMIT_EXCEEDED
            - INTERNAL_ERROR
            - PAYMENT_FAILED
        message:
          type: string
        details:
          nullable: true
        trace_id:
          type: string
...
    QuizDetail:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        subject:
          type: string
        board:
          type: string
        totalQuestions:
          type: integer
        timeLimit:
          type: integer
        questions:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              text:
                type: string
              type:
                type: string
              options:
                type: array
                items:
                  type: object
    CreateAttemptRequest:
      type: object
      properties:
        startedAt:
          type: string
          format: date-time
    CreateAttemptResponse:
      type: object
      properties:
        attemptId:
          type: integer
          format: int64
        expiresAt:
          type: string
          format: date-time
    SubmitAttemptRequest:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            type: object
            properties:
              questionId:
                type: integer
              selectedOptionId:
                type: integer
                nullable: true
        submittedAt:
          type: string
          format: date-time
    AttemptResult:
      type: object
      properties:
        attemptId:
          type: integer
        score:
          type: number
        correctCount:
          type: integer
        wrongCount:
          type: integer
        skippedCount:
          type: integer
        totalQuestions:
          type: integer
        timeTaken:
          type: integer
    ReviewDetail:
      type: object
      properties:
        score:
          type: number
        questions:
          type: array
          items:
            type: object
    ReportsPage:
      type: object
      properties:
        reports:
          type: array
          items:
            type: object
        total:
          type: integer
        page:
          type: integer
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
paths:
...
  /quizzes/{quizID}:
    get:
      summary: Get quiz detail (no answers)
      parameters:
        - name: quizID
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Quiz detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizDetail"
        "404":
          $ref: "#/components/responses/NotFound"
  /quizzes/{quizID}/attempts:
    post:
      summary: Start a quiz attempt
      parameters:
        - name: quizID
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAttemptRequest"
      responses:
        "201":
          description: Attempt started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateAttemptResponse"
  /quizzes/{quizID}/attempts/{attemptID}/submit:
    put:
      summary: Submit quiz attempt answers
      parameters:
        - name: quizID
          in: path
          required: true
          schema:
            type: integer
        - name: attemptID
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitAttemptRequest"
      responses:
        "200":
          description: Attempt scored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttemptResult"
        "409":
          description: Attempt not in submittable state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /quizzes/{quizID}/attempts/{attemptID}/review:
    get:
      summary: Review completed attempt
      parameters:
        - name: quizID
          in: path
          required: true
          schema:
            type: integer
        - name: attemptID
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Review detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewDetail"
  /users/me/reports:
    get:
      summary: Paginated quiz attempt reports
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Reports page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportsPage"
...